Add next improvements
