+++ src/openapi/transport/types.ts
@@ -1,7 +1,5 @@
 import type { StringTemplateArgs } from '../../utils/string';
 
-export type HTTPMethodType = 'get' | 'put' | 'post' | 'delete' | 'patch' | 'options' | 'head';
-
 export type HTTPStatusCode = 401 | 404 | 200 | 201 | 500;
 
 export type UseCloud = {
@@ -39,19 +37,3 @@ export interface TransportCoreOptions {
 
 // eslint-disable-next-line max-len
 export type HTTPMethodInputArgs = [string | undefined, string | undefined, StringTemplateArgs | undefined, TransportCoreOptions | undefined];
-
-export interface OAPICallResult {
-    response?: string | Blob | Record<string, unknown>;
-    status: number;
-    headers: Headers;
-    size: number;
-    url: string;
-    responseType?: string;
-    isNetworkError?: never;
-}
-
-export interface NetworkFailure {
-    message?: string | Error;
-    isNetworkError: true;
-    status?: never;
-}
